Print method for calibrate_thresholds objects

By default prints only the res_summary table from an object of class 'calibrate_thresholds'. The table can be limited to a range of type 1 error and a minimum value of power using the arguments 'type1_range' and 'minimum_power' respectively.

# S3 method for calibrate_thresholds
print(x, type1_range = c(0, 1), minimum_power = 0, ...)

Arguments

x

an object of class 'calibrate_thresholds', usually returned by the calibrate_thresholds function

type1_range

a vector specifying the minimum and maximum acceptable type I error. Specify c(0, 1) to return the full range of resulting type I error. Defaults to c(0, 1)

minimum_power

a numeric between 0 and 1 specifying the minimum acceptable power. Specify 0 to return the full range of resulting power. Defaults to 0.

...

ignored

Value

Returns a tibble

Examples

# Setting S = 50 and nsim = 50 for speed
# In practice you would want a much larger sample and more simulations

set.seed(123)

cal_tbl1 <- calibrate_thresholds(
  p_null = 0.1,
  p_alt = 0.4,
  n = seq(5, 15, 5),
  N = 15,
  pp_threshold = c(0.85, 0.9),
  ppp_threshold = c(0.1, 0.2),
  S = 50,
  nsim = 50
  )
#> Joining, by = c("sim_num", "pp_threshold", "ppp_threshold")

print(cal_tbl1)
#> # A tibble: 4 × 8
#>   pp_threshold ppp_threshold mean_n1_null prop_pos_null prop_stopped_null
#>          <dbl>         <dbl>        <dbl>         <dbl>             <dbl>
#> 1         0.85           0.1         12.1          0.1               0.42
#> 2         0.85           0.2          9.5          0.06              0.64
#> 3         0.9            0.1         10.8          0.04              0.5 
#> 4         0.9            0.2          8.9          0.04              0.74
#>   mean_n1_alt prop_pos_alt prop_stopped_alt
#>         <dbl>        <dbl>            <dbl>
#> 1        14.8         0.92             0.02
#> 2        14.8         0.92             0.02
#> 3        14.8         0.9              0.02
#> 4        14.7         0.88             0.04
print(cal_tbl1, type1_range = c(0.05, 0.1), minimum_power = 0.9)
#> # A tibble: 2 × 8
#>   pp_threshold ppp_threshold mean_n1_null prop_pos_null prop_stopped_null
#>          <dbl>         <dbl>        <dbl>         <dbl>             <dbl>
#> 1         0.85           0.1         12.1          0.1               0.42
#> 2         0.85           0.2          9.5          0.06              0.64
#>   mean_n1_alt prop_pos_alt prop_stopped_alt
#>         <dbl>        <dbl>            <dbl>
#> 1        14.8         0.92             0.02
#> 2        14.8         0.92             0.02